Made a script to download Xenia and used in a post-build event

+++ src/main.cpp
@@ -8,8 +8,8 @@ extern float g_DisplayHeight;
 void __cdecl main()
 {
     App app;
-    app.m_d3dpp.BackBufferWidth = g_DisplayWidth;
-    app.m_d3dpp.BackBufferHeight = g_DisplayHeight;
+    app.m_d3dpp.BackBufferWidth = static_cast<uint32_t>(g_DisplayWidth);
+    app.m_d3dpp.BackBufferHeight = static_cast<uint32_t>(g_DisplayHeight);
 
     // Make sure display gamma is correct.
     app.m_d3dpp.BackBufferFormat = static_cast<D3DFORMAT>(MAKESRGBFMT(D3DFMT_A8R8G8B8));
